chore: update build script to seperate patch buildings chore: reorganize patches
