+++ res/patch/vampire/dawnguard-vampire-lord-transform-fix/scripts/source/DLC1VampireTransformVisual.psc
